Telephony: Fix for MMI codes
Telephony: Fix for MMI codes for call forwarding

Fix for issue - The UI doesn't register for call forwarding when
*SC*SI# is dialed where SI is the number to which call should
be forwarded.

3GPP specification 22.030 sec. 6.5.2

The UE shall determine from the context whether,
an entry of a single *, activation or registration
was intended.

For example, a call forwarding request with a single *
would be interpreted as registration if containing a
forwarded-to number, or an activation if not.

Telephony: Enable MMI codes for CDMA Global devices operating in GSM/UMTS

Detects MMI codes for CustomerSupport, Voice Mail, CLIP and CLIR and
translates the service code to corresponding dialing numbers when the
CDMA Global phone is in GSM/UMTS networks.

Telephony: Handle COLP/COLR/CNAP MMI strings.

Currently COLP/COLR/CNAP interrogation is not supported and incorrectly
processed as USSD request. Changes done to reject the MMI command and
notify user regarding the unsupported command.

Change-Id: Idc80ce0797619b0ba30fafdc4d0726318d2fbecb