Add functionality to disable Watchdog killing System Server
Introduced a way to control the behavior of Android WatchDog for
any system process if its stuck.

This is needed for live debugging of user space issues.
Otherwise Android Watch Dog will kill the process that is
stuck for Monitor time.

Before watchdog kill system_server, we will get the tombstones of
system_server.

-Added logic to obtain the WDOG behavior based on System Properties.

-Now sends SIGSTOP once the System property is set.

Change-Id: I897cc07142710aba87608ae5fb94049127bdb1c6