frameworks/base:Bluetooth: List HID Mouse as Paired device
As per Bluetooth HID specification, Authentication is optional for
Mouse. During Bluetooth turn on process, The devices listed by Bluez
in 'Devices' property of AdapterProperties, will be listed under Paired
device if their 'Paired' property is true. For HID mouses, this property
would not be set, but it will be listed under 'Devices', if it was
previously used. The change lists the HID Mouse under paired device
list.

Change-Id: I8b31608b27911d7a48ff6458adff48b313e94319
CRs-fixed: 327359