libstagefright: Fix LPA mute issue via browser.
-When Playing LPA clip via HTML link, mute option
fails to work.
-In MediaPlayerService, setVolume API handles only
for non-LPA case. Need to change to call LPA Volume
update too.
-Call mSession setVolume incase of LPA. The AudioFlinger
has to keep track of previous volume when muted. This
volume is again applied back when unmuted since App
sents volume as unity when unmuted which is not the previous
volume before mute. This change fixes the below issues
--mute/unmute option via browser
--increase/decrease volume when mute-should not affect mute option.
--While in mute pause for 3sec and resume,mute is lost.

Change-Id: If546a453feb999989074e5cecdf6ec1f81c95e89
CRs-Fixed: 327159