Cne: Add FMC modules
Provides FMC solution where cellular data is tunneled over wifi on top
of the CT protocol. These modules provide the APIs for the application
to request start/stop FMC and routing the request to the data service.
It notifies the application for the FMC status reported from the data
service.

Change-Id: Iabeb88bb677b23223f9069cd9f1d87941a22698d