libstagefright: Crop the native window with display resolution
We provide the native window with the display width and height in
set_geometry.  On the surfaceflinger side, they assume that this is
aligned, which causes problems for HDMI.  Instead we'll now pass in the
aligned stride and slice in native_window_set_geometry, then specify
the desired display width and height and call native_window_set_crop.

Change-Id: Iecfe11ded743715ea2eabad4a3fddbc1b04b5005
CRs-Fixed: 332778