Bluetooth: Get socket ptr through get_socketData API in reconnect.
Async connect option for Bluetooth Socket has logic to reattempt
when the error is EALREADY. In this case the socket is released and
again allocated before retry connect option. Found the Socket ptr
is not properly updated in this usecase. The current change is to
have proper assignment for the variable being used.

Change-Id: I0b3ac91326adeb78d0ef97868c7fec5bcaa4d49e
CRs-Fixed: 330462