frameworks/base: Fix for wifi direct disable while suspend/resume.
When the Wi-Fi Direct is enabled after the Wi-Fi station is On the
station mode is TURNED OFF and Wi-FI Direct TURNED ON .After the
Wi-Fi Direct is ON and if a suspend and resume is done , the Wi-Fi
sta mode was getting activated and the Wi-Fi Direct was TURNED OFF.
The expectation here is that the Wi-Fi Direct should be TURNED ON
as we have entered the suspend state and resumed back in Wi-Fi
direct mode. To avoid this , a condition was added in the resume
path to check if Wi-Fi Direct is enabled and if yes the Wi-Fi
station mode is not enabled again.

Change-Id: I210a386913c0357c0000e3e2fd3643dafd346a8e
CRs-fixed: 332440