Bluetooth: Handling proper state change for HFP when HeadsetService restarts.
If the Phone App crashes for some reason, it recovers back and all
services will resart. But the shutdown and restart of BluetoothHeadsetService
during Headset connected state causes connection to be lost. And the connection
will not be reestablished as the BluetoothServices state variables are not
updated. The current change will detect Service disconnect and updates the
state varaiables properly.

Change-Id: Ib001da80418a70978fcf299cd7ba6c583a39c391