Bluetooth: Avoid call to getServiceChannelNative for device added in Scan.
During scan when devices are added, based on EIR response UUIDs, getService
ChannelNative is being called. Doing SDP is not needed for scanned devices. Anyhow
as the DBUS is busy due to scan, the synchronous API for getting service
records is timing out and the call is failing. This failure is causing
suspicious error log. The current change takes care of newly added devices.

CRs-Fixed: 327333

Change-Id: I6a149200ce95f9363a4c70b580d243f442103ab6