Bluetooth: ConnectOtherProfiles logic made robust for incoming pairing.
When remote devices like carkit have option to scan and connect to Phone,
connect other profiles logic is not being effective. The reason is
queing for outgoing connect to second profile happens only when device
priority is set to On. If Sdp initiated from DUT is not complete when
incoming connection happens, queuing connect other profiles wont happen.
The current logic take care of this miss.

CRs-Fixed: 337134

Change-Id: I712b05350789e309aca751f877fc0bda1426bc09