libstagefright:FIRST_FRAME is set with audioEOS event
As audioEOS event is posted, timesource is changed from audio clock to
system timesource.Because of the difference in two timeclock,it assumes it
to be late/early and leads to seek.
By setting the FIRST_FRAME with audioEOS event,the timesource difference is
taken into account while lateness calculation.

CRs-Fixed: 335557

Change-Id: I034f9f24d7e263a3f1ee150ff81cf9f094a8425d