libRS: Correct the screen freezing issue
This fixes a timing issue where it could calculate a delay of 0
(indicating wait forever) when it has no pending commands to actually
execute. In such cases, it should just break out of the playback loop.
This also fixes a small issue with returning whether or not to redraw.
(Provided by Stephen Hines <srhines@google.com>)

Change-Id: I645522720dcf98a9010de70a195cf2020d1dfbd0
CRs-Fixed: 330352