UsbDeviceManager: Add support to append an extra function in a composition
When "rndis,diag,adb" composition is enabled through setprop, composition
switch happens, but the Rndis fails to configure to get the IP.
It is required to set the USB Tethering option from the UI to configure
Rndis parameters, but the target switches to "rndis,adb" composition when
USB Tethering option is set.

This change allows the user to add the extra function using the following
command through adb shell in a new system property "sys.usb.config.extra"

For example: setprop sys.usb.config.extra diag

Thus allows the target to switch to "rndis,diag,adb" composition when diag
is enabled in set.usb.config.extra and UI Tethering option is set.

CRs-Fixed: 338602
Change-Id: Ie2ceaa72708f244559675accf34b5b71a0f3b22f
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>