Bluetooth: Even device is in sleep, Discoverable Mode toggles on timer expiry
Timer used for the Discoverable mode is turned off when device is
in Sleep, due to which device will be in Discoverable even after
the timer expires.
AlarmManager Service which will not get turned off in sleep is
used to wake up the device when timeout is occured for the
Discoverable Mode.

Change-Id: I8ed74615956bbebd521c2ace58626a42e8b0773e
CRs-fixed: 337994